Replacement refactoring
Replacements are broken down to individual calls lazily on first execution.
This is done by new ReplacementBlockNode, which executes RHS store to tmp
variable then the actual replacement and then it returns the RHS value.

ReplacementBlockNode creates the node for actual replacement depending on the
replacement structure: if all the calls are specials, it creates
ReplacementNodeSpecial, which takes advantage of the properties of specials.
Otherwise it creates regular replacement node that executes exactly the same
function calls as GnuR would.