Put errors log to user home if package installation is running
The working directory of package installation is in /tmp and gets deleted right
after the installation finishes even if not successful. This deletes the error
log. We do heuristic check if package installation is in progress and in such
case we do not use the working directory as location for errors log.